Splitted starHistory methods into a seperate file
